+++ Back/.metadata/version.ini
@@ -1,3 +1,3 @@
-#Mon May 26 18:13:41 KST 2025
+#Tue May 27 10:03:52 KST 2025
 org.eclipse.core.runtime=2
 org.eclipse.platform=4.29.0.v20230903-1000
