+++ Front/sodam/lib/login/login_page.dart
@@ -1,8 +1,20 @@
 import 'package:flutter/material.dart';
-import 'package:sodam/main_page.dart';
-import 'find_id_page.dart';
-import 'reset_password_page.dart';
-import 'guest_warning_page.dart';
+import '../main_page.dart';
+import 'package:dio/dio.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+
+final dio = Dio();
+
+void configureDio() {
+  // 기본 옵션 설정
+  dio.options.baseUrl = 'http://10.0.2.2:8080'; // 기본 URL
+  dio.options.connectTimeout = Duration(seconds: 5); // 연결 타임아웃: 5초
+  dio.options.receiveTimeout = Duration(seconds: 3); // 응답 수신 타임아웃: 3초
+  dio.options.headers = {
+    'Content-Type': 'application/json; charset=UTF-8',
+    'Authorization': 'Bearer YOUR_ACCESS_TOKEN', // 예시: 기본 인증 헤더
+  };
+}
 
 class LoginPage extends StatefulWidget {
   const LoginPage({super.key});
@@ -14,8 +26,16 @@ class LoginPage extends StatefulWidget {
 class _LoginPageState extends State<LoginPage> {
   final _idController = TextEditingController();
   final _pwController = TextEditingController();
+
+
   bool autoLogin = false;
   bool loginError = false;
+
+  @override
+  void initState() {
+    super.initState();
+    configureDio(); // 여기에 추가
+  }
 
   @override
   void dispose() {
@@ -26,7 +46,9 @@ class _LoginPageState extends State<LoginPage> {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
+    return Theme(
+        data: ThemeData.light().copyWith(brightness: Brightness.light),
+    child: Scaffold(
       backgroundColor: Colors.white,
       body: SafeArea(
         child: Padding(
@@ -75,42 +97,19 @@ class _LoginPageState extends State<LoginPage> {
 
                     // 자동로그인 체크박스
                     Row(
-                      mainAxisAlignment: MainAxisAlignment.end,
                       children: [
-                        GestureDetector(
-                          onTap: () {
-                            Navigator.push(
-                              context,
-                              MaterialPageRoute(builder: (_) => const FindIdPage()),
-                            );
-                          },
-                          child: const Text(
-                            '접속이름',
-                            style: TextStyle(
-                              decoration: TextDecoration.underline,
-                              color: Colors.black,
-                            ),
-                          ),
+                        Checkbox(
+                          value: autoLogin,
+                          onChanged: (val) => setState(() => autoLogin = val ?? false),
                         ),
-                        const Text(' / '),
-                        GestureDetector(
-                          onTap: () {
-                            Navigator.push(
-                              context,
-                              MaterialPageRoute(builder: (_) => const ResetPasswordPage()),
-                            );
-                          },
-                          child: const Text(
-                            '비밀번호 찾기',
-                            style: TextStyle(
-                              decoration: TextDecoration.underline,
-                              color: Colors.black,
-                            ),
-                          ),
+                        const Text('자동로그인'),
+                        const Spacer(),
+                        const Text(
+                          '접속이름 / 비밀번호 찾기',
+                          style: TextStyle(decoration: TextDecoration.underline),
                         ),
                       ],
                     ),
-
                   ],
                 ),
               ),
@@ -128,7 +127,7 @@ class _LoginPageState extends State<LoginPage> {
                   onPressed: () {
                     Navigator.pushReplacement(
                       context,
-                      MaterialPageRoute(builder: (_) => const GuestWarningPage()),
+                      MaterialPageRoute(builder: (_) => const MainPage()),
                     );
                   },
                   child: const Text('비회원 접속'),
@@ -146,21 +145,22 @@ class _LoginPageState extends State<LoginPage> {
                     padding: const EdgeInsets.symmetric(vertical: 16),
                     shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
                   ),
-                  onPressed: () {
-                    final id = _idController.text;
-                    final pw = _pwController.text;
+                    onPressed: () async {
+                      final id = _idController.text;
+                      final pw = _pwController.text;
 
-                    if (id == 'sodam' && pw == '1234') {
-                      Navigator.pushReplacement(
-                        context,
-                        MaterialPageRoute(builder: (_) => const MainPage()),
-                      );
-                    } else {
-                      setState(() {
-                        loginError = true;
-                      });
-                    }
-                  },
+                      final success = await tryLogin(id, pw);
+                      if (success) {
+                        Navigator.pushReplacement(
+                          context,
+                          MaterialPageRoute(builder: (_) => const MainPage()),
+                        );
+                      } else {
+                        setState(() {
+                          loginError = true;
+                        });
+                      }
+                    },
                   child: const Text(
                     '접속',
                     style: TextStyle(color: Colors.black),
@@ -171,6 +171,32 @@ class _LoginPageState extends State<LoginPage> {
           ),
         ),
       ),
+    ),
     );
+  }
+  Future<bool> tryLogin(String id, String password) async {
+    try {
+      final res = await dio.get(
+        '/member/login',
+        queryParameters: {
+          'id': id,
+          'password': password,
+        },
+      );
+
+      print('로그인 응답값: ${res.data}'); // 디버깅용
+
+      // 백엔드에서 1020이면 로그인 성공
+      if (res.data == 1020) {
+        final prefs = await SharedPreferences.getInstance();
+        await prefs.setString('loggedInId', id);
+        return true;
+      } else {
+        return false;
+      }
+    } catch (e) {
+      print('로그인 오류: $e');
+      return false;
+    }
   }
 }
