--- Front/sodam/lib/login/my_page.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class MyPage extends StatelessWidget {
-  const MyPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(child: Text('내방 페이지'));
-  }
-}
