+++ Front/sodam/lib/main_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'pages/chat_page.dart';
-import 'pages/game_page.dart';
-import 'pages/my_page_views/my_page.dart';
+import 'chat/chat_page.dart';
+import 'game/game_page.dart';
+import 'myPage/my_page.dart';
 
 class MainPage extends StatefulWidget {
   const MainPage({super.key});
