--- Front/sodam/lib/pages/auth_choice_page.dart
@@ -1,106 +1,0 @@
-import 'package:flutter/material.dart';
-import 'signup_page.dart';
-import 'login_page.dart';
-
-class AuthChoicePage extends StatelessWidget {
-  const AuthChoicePage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return
-      Theme(
-        data: ThemeData.light().copyWith(brightness: Brightness.light),
-    child: Scaffold(
-      backgroundColor: Colors.white, // 전체 배경 흰색
-      body: SafeArea(
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 30),
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              // 로고 이미지
-              Image.asset('assets/sodam_logo.png', width: 150), // 네 이미지 경로 맞게 조정!
-
-              const SizedBox(height: 40),
-
-              // 텍스트: 환영합니다.
-              const Text(
-                '환영합니다.',
-                style: TextStyle(
-                  fontSize: 24,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-
-              const SizedBox(height: 10),
-
-              const Text(
-                '완전! 무료 메시지와,\n게임을 부담없이 즐겨보세요!',
-                textAlign: TextAlign.center,
-                style: TextStyle(fontSize: 16),
-              ),
-
-              const SizedBox(height: 60),
-
-              // 가입 버튼
-              SizedBox(
-                width: double.infinity,
-                child: OutlinedButton(
-                  style: OutlinedButton.styleFrom(
-                    padding: const EdgeInsets.symmetric(vertical: 14),
-                    side: BorderSide(color: Colors.grey.shade300),
-                    shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(12)),
-                    backgroundColor: Colors.white,
-                  ),
-                  onPressed: () {
-                    Navigator.push(
-                      context,
-                      MaterialPageRoute(builder: (_) => const SignupPage()),
-                    );
-                  },
-                  child: const Text(
-                    '가입(데이터 필요)',
-                    style: TextStyle(
-                      fontSize: 16,
-                      color: Colors.black,
-                    ),
-                  ),
-                ),
-              ),
-
-              const SizedBox(height: 10),
-
-              // 접속 버튼
-              SizedBox(
-                width: double.infinity,
-                child: ElevatedButton(
-                  style: ElevatedButton.styleFrom(
-                    backgroundColor: const Color(0xFFC9DAB2),
-                    padding: const EdgeInsets.symmetric(vertical: 14),
-                    shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(12)),
-                  ),
-                  onPressed: () {
-                    Navigator.push(
-                      context,
-                      MaterialPageRoute(builder: (_) => const LoginPage()),
-                    );
-                  },
-                  child: const Text(
-                    '접속',
-                    style: TextStyle(
-                      fontSize: 16,
-                      color: Colors.black,
-                    ),
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    ),
-    );
-  }
-}
