--- Front/sodam/lib/pages/chat_page.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ChatPage extends StatelessWidget {
-  const ChatPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(child: Text('담소 페이지'));
-  }
-}
