--- Front/sodam/lib/pages/game_page.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class GamePage extends StatelessWidget {
-  const GamePage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(child: Text('놀이 페이지'));
-  }
-}
